add: dynamic inventory text

+++ src/bevy/world_gen.rs
@@ -14,7 +14,7 @@ pub fn spawn_random_world(mut commands: Commands, mut mma: MMA) {
 }
 
 impl PixelVariant {
-	fn weighted_natural_pool() -> (Vec<(PixelVariant, u8)>, u32) {
+	fn weighted_natural_pool() -> (Vec<(PixelVariant, u16)>, u32) {
 		let pool = PixelVariant::natural_pool();
 		let sum = pool.iter().map(|(_, natural)| natural.frequency as u32).sum();
 		(pool.into_iter().map(|(p, n)| (p, n.frequency)).collect(), sum)
